Refactor various map-specific behaviors/functions
