+++ src/elonacore.cpp
@@ -806,41 +806,6 @@ std::string txttargetlevel(int cc, int tc)
 
 
 
-std::string mapfile(int prm_365)
-{
-    if (area_data[prm_365].id == mdata_t::MapId::vernis)
-    {
-        return u8"vernis"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::port_kapul)
-    {
-        return u8"kapul"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::yowyn)
-    {
-        return u8"yowyn"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::derphy)
-    {
-        return u8"rogueden"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::palmia)
-    {
-        return u8"palmia"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::noyel)
-    {
-        return u8"noyel"s;
-    }
-    if (area_data[prm_365].id == mdata_t::MapId::lumiest)
-    {
-        return u8"lumiest"s;
-    }
-    return u8"vernis"s;
-}
-
-
-
 void initialize_picfood()
 {
     DIM3(picfood, 10, 9);
@@ -4767,8 +4732,7 @@ void removeworker(int map_id)
 
 void monster_respawn()
 {
-    if (area_data[game_data.current_map].id == mdata_t::MapId::museum
-        || area_data[game_data.current_map].id == mdata_t::MapId::shop)
+    if (area_data[game_data.current_map].is_museum_or_shop())
     {
         if (game_data.crowd_density < map_data.max_crowd_density / 2)
         {
@@ -5055,7 +5019,7 @@ TurnResult exit_map()
     }
     if (game_data.current_map != game_data.previous_map)
     {
-        if (map_data.is_town_or_guild()
+        if (map_is_town_or_guild()
             || game_data.previous_map == mdata_t::MapId::your_home
             || game_data.departure_date == 0)
         {
@@ -5263,231 +5227,6 @@ void save_map_local_data()
 
     // write data for items/character inventories local to this map
     ctrl_file(FileOperation2::map_items_write, u8"inv_"s + mid + u8".s2");
-}
-
-
-
-void map_proc_regen_and_update()
-{
-    if (game_data.date.hours() >= area_data[game_data.current_map]
-                                      .time_of_next_update_of_arena_random_seed)
-    {
-        area_data[game_data.current_map].arena_random_seed = rnd(10000);
-        area_data[game_data.current_map]
-            .time_of_next_update_of_arena_random_seed =
-            game_data.date.hours() + 24;
-    }
-    if (game_data.date.hours() >= map_data.next_regenerate_date)
-    {
-        if (map_data.should_regenerate == 0)
-        {
-            if (map_data.next_regenerate_date != 0)
-            {
-                for (int cnt = 0, cnt_end = (map_data.height); cnt < cnt_end;
-                     ++cnt)
-                {
-                    y = cnt;
-                    for (int cnt = 0, cnt_end = (map_data.width); cnt < cnt_end;
-                         ++cnt)
-                    {
-                        if (cell_data.at(cnt, y).feats != 0)
-                        {
-                            cell_featread(cnt, y);
-                            if (feat(1) >= 24 && feat(1) <= 28)
-                            {
-                                cell_featclear(cnt, y);
-                            }
-                        }
-                    }
-                }
-                for (const auto& cnt : items(-1))
-                {
-                    if (inv[cnt].number() == 0)
-                    {
-                        continue;
-                    }
-                    if (inv[cnt].id == 526)
-                    {
-                        if (inv[cnt].param1 < 10)
-                        {
-                            inv[cnt].param1 += 1;
-                            inv[cnt].image = 591;
-                            cell_refresh(
-                                inv[cnt].position.x, inv[cnt].position.y);
-                        }
-                    }
-                    if (map_data.is_town_or_guild())
-                    {
-                        if (inv[cnt].own_state < 0)
-                        {
-                            ++inv[cnt].own_state;
-                        }
-                        if (inv[cnt].own_state == 0)
-                        {
-                            inv[cnt].remove();
-                            cell_refresh(
-                                inv[cnt].position.x, inv[cnt].position.y);
-                        }
-                    }
-                }
-            }
-            for (auto&& cnt : cdata.others())
-            {
-                rc = cnt.index;
-                chara_clear_status_effects_b();
-                if (cnt.state() != Character::State::alive)
-                {
-                    continue;
-                }
-                if (cnt.is_temporary() == 1)
-                {
-                    if (rnd(2))
-                    {
-                        cnt.set_state(Character::State::empty);
-                        cell_data.at(cnt.position.x, cnt.position.y)
-                            .chara_index_plus_one = 0;
-                    }
-                }
-            }
-            p = rnd(map_data.height * map_data.width / 400 + 3);
-            if (map_data.type == mdata_t::MapType::world_map)
-            {
-                p = rnd(40);
-            }
-            if (map_data.type == mdata_t::MapType::town)
-            {
-                p = rnd(rnd(rnd(12) + 1) + 1);
-            }
-            if (map_data.type == mdata_t::MapType::guild)
-            {
-                p = rnd(p + 1);
-            }
-            for (int cnt = 0, cnt_end = (p); cnt < cnt_end; ++cnt)
-            {
-                map_randsite();
-            }
-        }
-        map_data.next_regenerate_date = game_data.date.hours() + 120;
-    }
-    if (game_data.date.hours() >= map_data.next_restock_date)
-    {
-        if (map_data.next_restock_date == 0)
-        {
-            renewmulti = 1;
-        }
-        else
-        {
-            renewmulti =
-                (game_data.date.hours() - map_data.next_restock_date) / 24 + 1;
-        }
-        if (area_data[game_data.current_map].id == mdata_t::MapId::ranch)
-        {
-            update_ranch();
-        }
-        if (area_data[game_data.current_map].id == mdata_t::MapId::your_home)
-        {
-            for (auto&& cnt : cdata.others())
-            {
-                if (cnt.state() != Character::State::alive)
-                {
-                    continue;
-                }
-                if (cnt.is_temporary() == 1)
-                {
-                    cnt.set_state(Character::State::empty);
-                    cell_data.at(cnt.position.x, cnt.position.y)
-                        .chara_index_plus_one = 0;
-                }
-            }
-        }
-        if (map_data.should_regenerate == 0)
-        {
-            for (int cnt = 0, cnt_end = (map_data.height); cnt < cnt_end; ++cnt)
-            {
-                y = cnt;
-                for (int cnt = 0, cnt_end = (map_data.width); cnt < cnt_end;
-                     ++cnt)
-                {
-                    if (cell_data.at(cnt, y).feats != 0)
-                    {
-                        cell_featread(cnt, y);
-                        x = cnt;
-                        if (feat(1) == 29)
-                        {
-                            for (int cnt = 0, cnt_end = (renewmulti);
-                                 cnt < cnt_end;
-                                 ++cnt)
-                            {
-                                if (feat >= tile_plant + 2)
-                                {
-                                    break;
-                                }
-                                grow_plant(
-                                    chipm(0, cell_data.at(x, y).chip_id_actual)
-                                            == 2
-                                        ? 1
-                                        : 0);
-                            }
-                            cell_featset(
-                                cnt, y, feat, feat(1), feat(2), feat(3));
-                        }
-                    }
-                }
-            }
-            if (map_data.is_town_or_guild()
-                || game_data.current_map == mdata_t::MapId::your_home)
-            {
-                for (auto&& cnt : cdata.others())
-                {
-                    if (cnt.state() != Character::State::alive)
-                    {
-                        continue;
-                    }
-                    generatemoney(cnt.index);
-                    if (cnt.id == 326)
-                    {
-                        if (itemfind(cnt.index, 60005, 1) == -1)
-                        {
-                            if (rnd(150) == 0)
-                            {
-                                flt();
-                                itemcreate(cnt.index, 707, -1, -1, 0);
-                            }
-                            else
-                            {
-                                flt(calcobjlv(cnt.level), calcfixlv());
-                                flttypeminor = 60005;
-                                itemcreate(cnt.index, 0, -1, -1, 0);
-                            }
-                        }
-                    }
-                    rc = cnt.index;
-                    if (rnd(5) == 0)
-                    {
-                        supply_new_equipment();
-                    }
-                    if (rnd(2) == 0)
-                    {
-                        if (inv_sum(rc) < 8)
-                        {
-                            flt(calcobjlv(cdata[rc].level),
-                                calcfixlv(Quality::bad));
-                            int stat = itemcreate(rc, 0, -1, -1, 0);
-                            if (stat != 0)
-                            {
-                                if (inv[ci].weight <= 0
-                                    || inv[ci].weight >= 4000)
-                                {
-                                    inv[ci].remove();
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        map_data.next_restock_date = game_data.date.hours() + 24;
-    }
 }
 
 
@@ -6640,582 +6379,6 @@ int place_random_nefias()
     }
     return p;
 }
-
-
-
-void map_proc_special_events()
-{
-    if (game_data.current_map == mdata_t::MapId::noyel)
-    {
-        if (game_data.released_fire_giant == 1)
-        {
-            if (cdata[game_data.fire_giant].state() == Character::State::alive)
-            {
-                if (game_data.crowd_density < 70)
-                {
-                    if (rnd(4) == 0)
-                    {
-                        flt();
-                        objlv = 1;
-                        chara_create(-1, 204, -3, 0);
-                    }
-                    if (rnd(10) == 0)
-                    {
-                        flt();
-                        objlv = 1;
-                        chara_create(-1, 185, -3, 0);
-                    }
-                }
-            }
-        }
-    }
-    if (game_data.executing_immediate_quest_type == 1008)
-    {
-        if (quest_data.immediate().progress != 3)
-        {
-            if (game_data.crowd_density
-                < game_data.left_minutes_of_executing_quest / 60)
-            {
-                dbid = 0;
-                if (rnd(4) == 0)
-                {
-                    dbid = 204;
-                }
-                if (rnd(10) == 0)
-                {
-                    dbid = 185;
-                }
-                if (dbid != 0)
-                {
-                    flt();
-                    objlv = 1;
-                    int stat = chara_create(-1, dbid, -3, 0);
-                    if (stat != 0)
-                    {
-                        cdata[rc].relationship = -1;
-                        cdata[rc].original_relationship = -1;
-                        cdata[rc].hate = 100;
-                        cdata[rc].enemy_id =
-                            quest_data.immediate().extra_info_2;
-                    }
-                }
-            }
-        }
-    }
-    if (game_data.current_map == mdata_t::MapId::tower_of_fire)
-    {
-        if (rnd(5) == 0)
-        {
-            r = sdata(50, 0) / 50;
-            if (r < 6)
-            {
-                dmg = (6 - r) * (6 - r) * 2;
-                txtef(3);
-                txt(i18n::s.get("core.locale.action.exit_map.it_is_hot"));
-                damage_hp(cdata[cc], dmg, -9);
-            }
-        }
-        return;
-    }
-    if (game_data.current_map == mdata_t::MapId::port_kapul)
-    {
-        if (game_data.current_dungeon_level == 25)
-        {
-            ++game_data.quest_flags.duration_of_kamikaze_attack;
-            x = 1;
-            y = rnd(map_data.height);
-            if (rnd(4) == 0)
-            {
-                x = map_data.width - 2;
-                y = rnd(map_data.height);
-            }
-            if (rnd(5) == 0)
-            {
-                x = rnd(map_data.width);
-                y = 1;
-            }
-            if (rnd(6) == 0)
-            {
-                x = rnd(map_data.width);
-                y = map_data.height - 2;
-            }
-            p = 237;
-            if (game_data.quest_flags.duration_of_kamikaze_attack > 50)
-            {
-                if (rnd(10) == 0)
-                {
-                    p = 245;
-                }
-            }
-            if (game_data.quest_flags.duration_of_kamikaze_attack > 100)
-            {
-                if (rnd(10) == 0)
-                {
-                    p = 244;
-                }
-            }
-            if (game_data.quest_flags.duration_of_kamikaze_attack > 150)
-            {
-                if (rnd(10) == 0)
-                {
-                    p = 244;
-                }
-            }
-            if (game_data.quest_flags.duration_of_kamikaze_attack == 250)
-            {
-                quest_update_journal_msg();
-                game_data.quest_flags.kamikaze_attack = 3;
-                txtef(9);
-                txt(i18n::s.get(
-                    "core.locale.misc.quest.kamikaze_attack.message"));
-                txt(i18n::s.get(
-                    "core.locale.misc.quest.kamikaze_attack.stairs_appear"));
-                cell_featset(18, 9, tile_downstairs, 11, 1);
-            }
-            flt();
-            chara_create(-1, p, x, y);
-            cdata[rc].hate = 1000;
-            cdata[rc].enemy_id = 0;
-            return;
-        }
-    }
-    if (game_data.current_map == mdata_t::MapId::jail)
-    {
-        if (rnd(50) == 0)
-        {
-            if (cdata.player().karma < -30)
-            {
-                txt(i18n::s.get("core.locale.misc.map.jail.repent_of_sin"));
-                modify_karma(cdata.player(), 1);
-                p = rnd(8) + 10;
-                if (sdata.get(p, 0).original_level >= 10)
-                {
-                    chara_gain_fixed_skill_exp(cdata.player(), p, -300);
-                }
-            }
-        }
-        if (cdata.player().karma >= -30)
-        {
-            if (rnd(1000) == 0)
-            {
-                cell_featread(29, 6);
-                if (feat(1) == 21)
-                {
-                    txt(i18n::s.get(
-                        "core.locale.misc.map.jail.guard_approaches"));
-                    txt(i18n::s.get("core.locale.misc.map.jail.leave_here"));
-                    txt(i18n::s.get(
-                        "core.locale.misc.map.jail.unlocks_your_cell"));
-                    cell_featset(29, 6, tile_dooropen, 20, 0, -1);
-                    snd(48);
-                }
-            }
-        }
-    }
-    if (game_data.current_map == mdata_t::MapId::shelter_)
-    {
-        if (game_data.weather == 2 || game_data.weather == 4
-            || game_data.weather == 1)
-        {
-            if (cdata.player().nutrition < 5000)
-            {
-                if (cdata.player().has_anorexia() == 0)
-                {
-                    snd(18);
-                    txt(i18n::s.get(
-                        "core.locale.misc.map.shelter.eat_stored_food"));
-                    cdata[cc].nutrition += 5000;
-                    show_eating_message();
-                }
-            }
-            if (game_data.continuous_active_hours >= 15)
-            {
-                game_data.continuous_active_hours = 13;
-            }
-            map_data.turn_cost = 1000000;
-        }
-        else if (map_data.turn_cost == 1000000)
-        {
-            map_data.turn_cost = 10000;
-            for (auto&& cnt : cdata.all())
-            {
-                cnt.turn_cost = 0;
-            }
-            txt(i18n::s.get(
-                "core.locale.misc.map.shelter.no_longer_need_to_stay"));
-        }
-    }
-    if (area_data[game_data.current_map].id == mdata_t::MapId::museum)
-    {
-        if (game_data.crowd_density > 0)
-        {
-            if (rnd(25) == 0)
-            {
-                txtef(9);
-                txt(i18n::s.get("core.locale.misc.map.museum.chats"));
-            }
-            if (rnd(25) == 0)
-            {
-                txtef(9);
-                txt(i18n::s.get("core.locale.misc.map.museum.chats2"));
-            }
-            if (rnd(15) == 0)
-            {
-                txtef(9);
-                for (int cnt = 0; cnt < 1; ++cnt)
-                {
-                    if (en)
-                    {
-                        break;
-                    }
-                    if (game_data.ranks.at(3) > 8000)
-                    {
-                        txt(u8"「退屈ぅー」"s,
-                            u8"「あまり見るものがないな」"s,
-                            u8"「こんなので見物料とるの？」"s,
-                            u8"館内は少し寂しい…"s);
-                        break;
-                    }
-                    if (game_data.ranks.at(3) > 5000)
-                    {
-                        txt(u8"「いいんじゃない〜」"s,
-                            u8"「まあ、普通の博物館だ」"s,
-                            u8"「恋人を連れてくればよかったかも」"s,
-                            u8"まあまあの客足だ。"s);
-                        break;
-                    }
-                    if (game_data.ranks.at(3) > 2500)
-                    {
-                        txt(u8"「この雰囲気好きだなぁ」"s,
-                            u8"「もう一度来ようよ」"s,
-                            u8"「時間が経つのを忘れるね！」"s,
-                            u8"館内はなかなか賑わっている。"s);
-                        break;
-                    }
-                    if (game_data.ranks.at(3) > 500)
-                    {
-                        txt(u8"「来て良かった♪」"s,
-                            u8"「よくこんなに集めたなあ」"s,
-                            u8"「むぅ…興味深い」"s,
-                            u8"客足が全く絶えない盛況ぶりだ。"s);
-                        break;
-                    }
-                    txt(u8"「素晴らしいコレクションだ！」"s,
-                        u8"「感動した」"s,
-                        u8"「帰りたくないわ♪」"s,
-                        u8"来客は食い入るように展示物を眺めている。"s);
-                }
-            }
-            return;
-        }
-    }
-    if (area_data[game_data.current_map].id == mdata_t::MapId::shop)
-    {
-        if (game_data.crowd_density > 0)
-        {
-            if (rnd(25) == 0)
-            {
-                txtef(9);
-                txt(i18n::s.get("core.locale.misc.map.shop.chats"));
-            }
-            return;
-        }
-    }
-}
-
-
-
-void map_reload_noyel()
-{
-    for (const auto& cnt : items(-1))
-    {
-        if (inv[cnt].id == 555 || inv[cnt].id == 600)
-        {
-            continue;
-        }
-        inv[cnt].remove();
-
-        cell_refresh(inv[cnt].position.x, inv[cnt].position.y);
-    }
-    if (area_data[game_data.current_map].christmas_festival == 1)
-    {
-        flt();
-        int stat = itemcreate(-1, 763, 29, 16, 0);
-        if (stat != 0)
-        {
-            inv[ci].own_state = 1;
-        }
-        {
-            flt();
-            int stat = itemcreate(-1, 686, 29, 16, 0);
-            if (stat != 0)
-            {
-                inv[ci].own_state = 1;
-            }
-        }
-        {
-            flt();
-            int stat = itemcreate(-1, 171, 29, 17, 0);
-            if (stat != 0)
-            {
-                inv[ci].param1 = 6;
-                inv[ci].own_state = 1;
-            }
-        }
-        {
-            flt();
-            int stat = itemcreate(-1, 756, 29, 17, 0);
-            if (stat != 0)
-            {
-                inv[ci].own_state = 5;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 345, 48, 19);
-            if (stat != 0)
-            {
-                cdata[rc].character_role = 3;
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 348, 30, 17);
-            if (stat != 0)
-            {
-                cdata[rc].character_role = 3;
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 174, 38, 19);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-                cdata[rc].is_hung_on_sand_bag() = true;
-                cdatan(0, rc) = i18n::s.get("core.locale.chara.job.fanatic");
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 347, 35, 19);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 347, 37, 18);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 347, 37, 21);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 347, 39, 20);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 347, 38, 21);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 70, 17, 8);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].only_christmas() = true;
-                cdata[rc].character_role = 1002;
-                cdata[rc].shop_rank = 10;
-                cdatan(0, rc) = snfood(cdatan(0, rc));
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 239, 25, 8);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].relationship = 0;
-                cdata[rc].original_relationship = 0;
-                cdata[rc].only_christmas() = true;
-                cdata[rc].character_role = 1018;
-                cdata[rc].shop_rank = 30;
-                cdatan(0, rc) = randomname();
-                cdatan(0, rc) = i18n::s.get(
-                    "core.locale.chara.job.souvenir_vendor", cdatan(0, rc));
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 271, 24, 22);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].relationship = 0;
-                cdata[rc].original_relationship = 0;
-                cdata[rc].only_christmas() = true;
-                cdata[rc].character_role = 1018;
-                cdata[rc].shop_rank = 30;
-                cdatan(0, rc) = randomname();
-                cdatan(0, rc) = i18n::s.get(
-                    "core.locale.chara.job.souvenir_vendor", cdatan(0, rc));
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 1, 38, 12);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].character_role = 1007;
-                cdata[rc].shop_rank = 10;
-                cdatan(0, rc) = snblack(cdatan(0, rc));
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 271, 28, 9);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].relationship = 0;
-                cdata[rc].original_relationship = 0;
-                cdata[rc].only_christmas() = true;
-                cdata[rc].character_role = 1022;
-                cdata[rc].shop_rank = 30;
-                cdatan(0, rc) = randomname();
-                cdatan(0, rc) = i18n::s.get(
-                    "core.locale.chara.job.street_vendor", cdatan(0, rc));
-            }
-        }
-        {
-            flt();
-            int stat = chara_create(-1, 271, 29, 24);
-            if (stat != 0)
-            {
-                cdata[rc].ai_calm = 3;
-                cdata[rc].relationship = 0;
-                cdata[rc].original_relationship = 0;
-                cdata[rc].only_christmas() = true;
-                cdata[rc].character_role = 1022;
-                cdata[rc].shop_rank = 30;
-                cdatan(0, rc) = randomname();
-                cdatan(0, rc) = i18n::s.get(
-                    "core.locale.chara.job.street_vendor2", cdatan(0, rc));
-            }
-        }
-        for (int cnt = 0; cnt < 20; ++cnt)
-        {
-            flt();
-            int stat = chara_create(-1, 349, -3, 0);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-            flt();
-            {
-                int stat = chara_create(-1, 350, -3, 0);
-                if (stat != 0)
-                {
-                    cdata[rc].only_christmas() = true;
-                }
-            }
-        }
-        for (int cnt = 0; cnt < 15; ++cnt)
-        {
-            flt();
-            int stat = chara_create(-1, 326, -3, 0);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-        }
-        for (int cnt = 0; cnt < 7; ++cnt)
-        {
-            flt();
-            int stat = chara_create(-1, 335, -3, 0);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-            {
-                flt();
-                int stat = chara_create(-1, 185, -3, 0);
-                if (stat != 0)
-                {
-                    cdata[rc].only_christmas() = true;
-                }
-            }
-            {
-                flt();
-                int stat = chara_create(-1, 274, -3, 0);
-                if (stat != 0)
-                {
-                    cdata[rc].only_christmas() = true;
-                }
-            }
-            {
-                flt();
-                int stat = chara_create(-1, 174, -3, 0);
-                if (stat != 0)
-                {
-                    cdata[rc].only_christmas() = true;
-                }
-            }
-        }
-        for (int cnt = 0; cnt < 3; ++cnt)
-        {
-            flt();
-            int stat = chara_create(-1, 332, -3, 0);
-            if (stat != 0)
-            {
-                cdata[rc].only_christmas() = true;
-            }
-            {
-                flt();
-                int stat = chara_create(-1, 185, -3, 0);
-                if (stat != 0)
-                {
-                    cdata[rc].only_christmas() = true;
-                }
-            }
-        }
-    }
-    else
-    {
-        for (auto&& cnt : cdata.others())
-        {
-            if (cnt.only_christmas() == 1)
-            {
-                chara_vanquish(cnt.index);
-            }
-        }
-    }
-}
-
-
 
 int random_material(int prm_1028, int prm_1029)
 {
@@ -8838,17 +8001,13 @@ void try_to_return()
             continue;
         }
         f = 0;
-        if (area_data[i].id == mdata_t::MapId::lesimas
-            || area_data[i].id == mdata_t::MapId::larna
-            || area_data[i].id == mdata_t::MapId::mansion_of_younger_sister
-            || area_data[i].id == mdata_t::MapId::the_void)
+        if (area_data[i].can_return_to())
         {
             f = 1;
         }
         if (game_data.wizard)
         {
-            if (area_data[i].type == mdata_t::MapType::town
-                || area_data[i].type == mdata_t::MapType::guild)
+            if (area_data[i].can_return_to_if_wizard())
             {
                 f = 1;
             }
@@ -14460,8 +13619,7 @@ TurnResult do_plant()
 {
     screenupdate = -1;
     update_screen();
-    if (map_data.type == mdata_t::MapType::world_map
-        || map_data.is_town_or_guild())
+    if (map_data.type == mdata_t::MapType::world_map || map_is_town_or_guild())
     {
         txt(i18n::s.get("core.locale.action.plant.cannot_plant_it_here"));
         update_screen();
@@ -15395,12 +14553,9 @@ void weather_changes()
             update_screen();
         }
     }
-    if (game_data.current_map != mdata_t::MapId::shelter_)
+    if (!map_prevents_random_events())
     {
-        if (game_data.current_map != mdata_t::MapId::jail)
-        {
-            proc_random_event();
-        }
+        proc_random_event();
     }
 }
