+++ src/initialize_map.cpp
@@ -130,11 +130,6 @@ static void _init_map_data()
     map_data.indoors_flag = area_data[game_data.current_map].is_indoor ? 1 : 2;
 }
 
-static bool _should_reveal_map()
-{
-    return dbg_revealmap || map_data.should_reveal_map();
-}
-
 static void _reveal_map()
 {
     for (int cnt = 0, cnt_end = (map_data.height); cnt < cnt_end; ++cnt)
@@ -151,7 +146,7 @@ static void _reveal_map()
 static void _init_map_visibility()
 {
     map_setfog();
-    if (_should_reveal_map())
+    if (map_should_reveal_fog())
     {
         _reveal_map();
     }
@@ -524,7 +519,7 @@ static void _refresh_map_character_other(Character& chara)
             }
         }
     }
-    if (map_data.is_town_or_guild())
+    if (map_is_town_or_guild())
     {
         chara.sleep = 0;
         if (game_data.date.hour >= 22 || game_data.date.hour < 7)
@@ -1075,8 +1070,7 @@ static void _proc_map_hooks_2()
     {
         _update_quest_escorts();
     }
-    if (area_data[game_data.current_map].id == mdata_t::MapId::museum
-        || area_data[game_data.current_map].id == mdata_t::MapId::shop)
+    if (area_data[game_data.current_map].is_museum_or_shop())
     {
         _spawn_museum_or_shop_crowds();
     }
@@ -1112,7 +1106,7 @@ static void _proc_map_hooks_2()
     {
         maybe_show_ex_help(14);
     }
-    if (map_data.is_town_or_guild()
+    if (map_is_town_or_guild()
         || game_data.current_map == mdata_t::MapId::your_home)
     {
         if (game_data.distance_between_town >= 16)
