+++ src/shop.hpp
@@ -6,7 +6,6 @@ namespace elona
 void shop_refresh_on_talk();
 void shop_refresh();
 void calc_number_of_items_sold_at_shop();
-void calc_trade_goods_price();
 void shop_load_shoptmp();
 void shop_sell_item();
